requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
feedparser==6.0.11
requests==2.32.3
newspaper3k==0.2.8
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.12
